fix: e2e upload avatar test - avoid remounting user profile fields
